Fix FrontiersmanZombie DM Spawner
Turns out it just needed the flag to get added to the non-replacement table; the command defaults to replacements only, and that table didn't exist so it was never being added.  Technically renaming the table will still work, but this way it won't attempt to spawn its own thing on top of any other mod also spawning on top of the dummy spawners.

A replacement table would also be valid, but the main part is that the entry wasn't finding the correct table to be added into.